Revert "Feature/authentication and data base"
