+++ .gitignore
@@ -4,4 +4,3 @@ npm-debug.log
 .nyc
 .env
 .DS_Store
-.rft
