+++ server.js
@@ -4,7 +4,7 @@
  * Module dependencies.
  */
 
-var app = require('./server/config/app');
+var app = require('./app');
 var debug = require('debug')('portfolio:server');
 var http = require('http');
