--- server/config/app.js
@@ -1,93 +1,0 @@
-var createError = require('http-errors');
-var express = require('express');
-var path = require('path');
-var cookieParser = require('cookie-parser');
-var logger = require('morgan');
-
-// modules for authentication
-let session = require('express-session');
-let passport = require('passport');
-let passportLocal = require('passport-local');
-let localStrategy = passportLocal.Strategy;
-let flash = require('connect-flash');
-
-
-
-// database setup
-let mongoose = require('mongoose');
-let DB = require('./db')
-
-// connect mongoose 
-mongoose.connect(DB.URI, {useNewUrlParser: true, useUnifiedTopology: true})
-let mongoDB = mongoose.connection;
-mongoDB.on('error', console.error.bind(console, "Connection Error: "))
-mongoDB.once('open', ()=>{
-  console.log('Connected to mongodb ...')
-})
-
-
-var indexRouter = require('../routes/index');
-var contactRouter = require('../routes/contact');
-var userRouter = require('../routes/user');
-
-var app = express();
-
-// view engine setup
-app.set('views', path.join(__dirname, '../views'));
-app.set('view engine', 'ejs');
-
-app.use(logger('dev'));
-app.use(express.json());
-app.use(express.urlencoded({ extended: false }));
-app.use(cookieParser());
-app.use(express.static(path.join(__dirname, '../../public')));
-app.use(express.static(path.join(__dirname, '../../node_modules')))
-
-//setup express session
-app.use(session({
-  secret: "SomeSecret",
-  saveUninitialized: false,
-  resave: false
-}))
-
-// Initialize flash
-app.use(flash());
-
-//initialize passport
-app.use(passport.initialize());
-app.use(passport.session());
-
-// passport user configuration
-// create a Model instance
-let userModel = require('../models/user');
-let user = userModel.User;
-
-// serialize and deserialized info
-passport.serializeUser(userModel.User.serializeUser());
-passport.deserializeUser(userModel.User.deserializeUser());
-
-app.use('/', indexRouter);
-app.use('/user', userRouter)
-app.use('/contacts', contactRouter);
-
-// implement a User Authentication Strategy
-passport.use(user.createStrategy());
-
-
-// catch 404 and forward to error handler
-app.use(function(req, res, next) {
-  next(createError(404));
-});
-
-// error handler
-app.use(function(err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render('error', {title: 'Error'});
-});
-
-module.exports = app;
