--- server/config/db.js
@@ -1,3 +1,0 @@
-module.exports = {
-    "URI": "mongodb://localhost/business_contacts" 
-}
