--- server/controllers/index.js
@@ -1,135 +1,0 @@
-const { urlencoded } = require('express');
-let express = require('express');
-let router = express.Router();
-let mongoose = require('mongoose');
-let passport = require('passport');
-
-let userModel = require('../models/user');
-let User = userModel.User;
-
-module.exports.displayHomePage = (req, res, next) => {
-    res.render('index', {title: 'Home', displayName: req.user ? req.user.displayName : ''});
-}
-
-module.exports.displayAboutPage = (req, res, next) => {
-    res.render('index', { title: 'About me', displayName: req.user ? req.user.displayName : ''});
-}
-
-module.exports.displayProjectsPage = (req, res, next) => {
-    res.render('index', { title: 'Projects', displayName: req.user ? req.user.displayName : ''});
-}
-
-module.exports.displayServicesPage = (req, res, next) => {
-    res.render('index', { title: 'Services', displayName: req.user ? req.user.displayName : ''});
-}
-
-module.exports.displayContactPage = (req, res, next) => {
-    res.render('index', { title: 'Contact', displayName: req.user ? req.user.displayName : ''});
-}
-
-module.exports.displayLoginPage = (req, res, next) => {
-    // check if the user is already logged in
-    if(!req.user)
-    {
-        res.render('auth/login', 
-        {
-           title: "Login",
-           messages: req.flash('loginMessage'),
-           displayName: req.user ? req.user.displayName : '' 
-        })
-    }
-    else
-    {
-        return res.redirect('/');
-    }
-}
-
-module.exports.performLogin = (req, res, next) => {
-    passport.authenticate('local', 
-    (err, user, info) => 
-    {
-        if(err)
-        {
-            return next(err)
-        }
-        if(!user)
-        {
-            req.flash('loginMessage', 'Authentication Error');
-            return res.redirect('/login')
-        }
-        req.login(user, (err) => {
-            if(err)
-            {
-                return next(err); 
-            }
-            return res.redirect('/contacts');
-        });
-    })(req, res, next);
-}
-
-module.exports.displayRegisterPage = (req, res, next) => {
-    // check if the user is not already logged in
-    if(!req.user)
-    {
-        res.render('auth/register',
-        {
-            title: 'Register',
-            messages: req.flash('registerMessage'),
-            displayName: req.user ? req.user.displayName : ''
-        });
-    }
-    else
-    {
-        return res.redirect('/');
-    }
-}
-
-module.exports.processRegisterPage = (req, res, next) => {
-    // instantiate a user object
-    let newUser = new User({
-        username: req.body.username,
-        //password: req.body.password
-        email: req.body.email,
-        displayName: req.body.displayName
-    });
-
-    User.register(newUser, req.body.password, (err) => {
-        if(err)
-        {
-            console.log("Error: Inserting New User");
-            if(err.name == "UserExistsError")
-            {
-                req.flash(
-                    'registerMessage',
-                    'Registration Error: User Already Exists!'
-                );
-                console.log('Error: User Already Exists!')
-            }
-            return res.render('auth/register',
-            {
-                title: 'Register',
-                messages: req.flash('registerMessage'),
-                displayName: req.user ? req.user.displayName : ''
-            });
-        }
-        else
-        {
-            // if no error exists, then registration is successful
-
-            // redirect the user and authenticate them
-
-            /* TODO - Getting Ready to convert to API
-            res.json({success: true, msg: 'User Registered Successfully!'});
-            */
-
-            return passport.authenticate('local')(req, res, () => {
-                res.redirect('/contacts')
-            });
-        }
-    });
-}
-
-module.exports.performLogout = (req, res, next) => {
-    req.logout()
-    res.redirect('/')
-}
