--- server/views/auth/login.ejs
@@ -1,48 +1,0 @@
-<%- include ../partials/header.ejs %>  
-<%- include ../partials/main_nav.ejs %> 
-  <main class="container pt-5">
-    <% if(messages.length > 0) {%>
-      <div class="alert alert-danger">
-          <%= messages %>
-      </div>
-    <% } %>
-    <div class="row d-flex justify-content-center align-items-center h-100">
-      <div class="col col-xl-5">
-        <div class="card" style="border-radius: 1rem; background-color: #cec7c8;">
-          <div class="row g-0">
-            <div class="col-md-12 col-lg-12 d-flex align-items-center">
-              <div class="card-body p-4 p-lg-5 text-black">
-                <form action="" method="post" autocomplete="off">
-                  <div class="d-flex align-items-center mb-3 pb-1">
-                    <span class="h1 fw-bold mb-0" style="color: #6e6b6c">Sign in</span>
-                    <i class="fas fa-sign-in-alt fa-2x px-3" style="color: #8d898a;"></i>
-                  </div>
-                  <fieldset class="form-outline mb-4">
-                    <input type="text" id="username" name="username" class="form-control form-control-lg" required autofocus/>
-                    <label class="form-label" for="username" style="color: #6e6b6c">Username</label>
-                  </fieldset>
-
-                  <fieldset class="form-outline mb-4">
-                    <input type="password" id="password" name="password" class="form-control form-control-lg" required/>
-                    <label class="form-label" for="password" style="color: #6e6b6c">Password</label>
-                    or
-                    <a href="/register">Register Here</a>
-                  </fieldset>
-
-                  <div class="pt-1 mb-4">
-                    <input type='submit' value='Login' class="btn btn-lg btn-block" style="background-color:#ada9a9; color: #eeeced" type="button"></input>
-                    <a href="/">
-                      <input type='submit' value='Cancel' class="btn btn-lg btn-block" style="background-color:#ada9a9; color: #eeeced" type="button"></input>
-                    </a>
-                  </div>
-                </form>
-
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </main>
-  <%- include ../partials/bottom_nav.ejs %>   
-  <%- include ../partials/footer.ejs %>
