+++ views/partials/main_nav.ejs
@@ -38,26 +38,6 @@
             <% } %>
             <i class="fas fa-comments-dollar"></i> Contact Me</a>
           </li>
-          <li class="nav-item">
-            <% if (title == "Contact List") { %>
-            <a class="nav-link active" href="/contacts">
-            <% } else { %>
-            <a class="nav-link" href="/contacts">
-            <% } %>
-            <i class="fas fa-address-book"></i>Contact List</a>
-          </li>
-
-          <li class="nav-item">
-          <% if(typeof displayName == 'string' && displayName) { %>
-          <li class="navbar-text"><i class="fas fa-user"></i> Welcome, <%= displayName %></li>
-          <li class="nav-item"><a href="/logout" class="nav-link">
-            <i class="fas fa-sign-out-alt"></i> Logout</a></li>
-          <% } else { %>
-          <li class="nav-item"><a href="/login" class="nav-link">
-            <i class="fas fa-sign-in-alt"></i> Login</a></li>
-          <% } %>
-          </li>
-          
         </ul>
       </div>
     </div>
